# Optimize parsing/conversion of TextDecorations from string, reduce allocations